requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
multidict==6.1.0
numpy==2.0.2
propcache==0.3.0
pyahocorasick==2.1.0
RapidFuzz==3.12.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
